@@ -1,8 +1,8 @@
 package com.example.android.recipes.adapters;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -12,7 +12,6 @@ import android.widget.TextView;
 import com.bumptech.glide.Glide;
 import com.example.android.recipes.R;
 import com.example.android.recipes.models.network.Recipe;
-import com.example.android.recipes.ui.RecipesListActivity;
 
 import java.util.ArrayList;
 import java.util.List;
