@@ -4,7 +4,7 @@ package com.example.android.recipes.models.network;
  * This class is used to help move all the variables via intents
  */
 
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 public class RecipeSelectedManager {
 
