@@ -1,17 +1,15 @@
 package db;
 
 
-import android.arch.lifecycle.LiveData;
-import android.arch.persistence.room.Delete;
-import android.arch.persistence.room.Insert;
-import android.arch.persistence.room.Query;
-import android.arch.persistence.room.Update;
-
-import com.example.android.recipes.models.network.Recipe;
+import androidx.lifecycle.LiveData;
+import androidx.room.Delete;
+import androidx.room.Insert;
+import androidx.room.Query;
+import androidx.room.Update;
 
 import java.util.List;
 
-@android.arch.persistence.room.Dao
+@androidx.room.Dao
 public interface Dao {
 
     @Insert
